refactor: move the node-compat parsing from Wrangler into Miniflare
This logic is now also needed inside Miniflare to determine how to handle node imports. So this commit moves the relevant code in there and references it from Wrangler.

The validation of the config is kept in Wrangler, since Miniflare doesn't need that
particularly and the validation relies upon Wrangler logger and UserError.
