+++ packages/wrangler/src/pages/buildFunctions.ts
@@ -10,9 +10,9 @@ import { writeRoutesModule } from "./functions/routes";
 import { convertRoutesToRoutesJSONSpec } from "./functions/routes-transformation";
 import { getPagesTmpDir, RUNNING_BUILDERS } from "./utils";
 import type { BundleResult } from "../deployment-bundle/bundle";
-import type { NodeJSCompatMode } from "../deployment-bundle/node-compat";
 import type { PagesBuildArgs } from "./build";
 import type { Config } from "./functions/routes";
+import type { NodeJSCompatMode } from "miniflare";
 
 /**
  * Builds a Functions worker based on the functions directory, with filepath and handler based routing.
